Annotations/FrameSelector_DesktopGUI/requirements.txt
opencv-python==4.8.1.78
boto3==1.34.11
python-dotenv==1.0.0
pillow==10.1.0
